Assignment04/Assigment 00 to 09/qr_code_decode_generator/pyproject.toml
[project]
name = "qr-code-decode-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pyzbar>=0.1.9",
    "qrcode>=8.1",
]
